UPDATE refactor of the project structure
